license_portal/requirements.txt
Django==3.2
psycopg2
time_machine
djangorestframework
